requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.4.2
python-dotenv==1.0.0
certifi==2024.2.2
dnspython==2.4.2 